Some refactors.
